\name{rsfse-package}
\alias{rsfse-package}
\alias{rsfse}
\docType{package}
\title{
Random Survival Forest with Space Extensions for Censored Data
}
\description{
Random Survival Forest with Space Extensions for Censored Data
}
\details{
\tabular{ll}{
Package: \tab rsfse\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-03-23\cr
License: \tab G\cr
}
Random Survival Forest with Space Extensions for Censored Data
}
\author{
Hong wang

Maintainer: Hong Wang <wh@csu.edu.cn>

}
\references{
Random Survival Forest with Space Extensions for Censored Data, submitted to Artificial Intelligence in Medicine
}
\keyword{ package }
\seealso{

}
\examples{

}
